refactor: moved `messages` into "modules" folder
