+++ src/front/modules/messages/components/ChatConversation.jsx
