+++ src/front/modules/messages/components/ChatSidebar.jsx
