+++ src/front/modules/messages/components/Conversation-subcomponents/MessageBubble.jsx
