+++ src/front/modules/messages/hooks/useIntersectionObserver.js
