+++ src/front/modules/messages/utils/formatTime.js
