+++ src/front/modules/messages/utils/normalize.js
